Add a Scheduler that determines what data to import
Affiliate Window’s data isn’t append-only. The
transactions are “confirmed” or “rejected” many
weeks after they appear in the API. That means we
need to repeatedly fetch the last N days of data
and update our database.

This commit introduces a scheduler that will be
responsible for figuring out what data to extract
to keep the model up to date. Currently, it
naively pulls in the last N days of data, but
it’ll need to look at “pending” transactions that
precede this window as well.